# backend/openapi.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "WattPrint API",
    "summary": "Household energy disaggregation (NILM) for WattPrint",
    "description": "Whole-house power in, appliance-level consumption out.\n\n* **Demo** — one household (Plegma House 101), every result precomputed: no model runs, every\n  call answers in milliseconds. Start with `GET /api/v1/demo/household`.\n* **Platform** — devices, raw readings, CSV import of NILM results and live disaggregation.\n\nErrors follow RFC 9457 (`application/problem+json`). Times are ISO 8601, UTC.\n",
    "version": "0.2.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "The host serving this document",
      "name": "current"
    }
  ],
  "paths": {
...
    "/api/v1/households/import": {
      "post": {
        "tags": [
          "disaggregation"
        ],
        "summary": "Import Disaggregation",
        "operationId": "importHouseholdDisaggregation",
        "parameters": [
          {
            "name": "household",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Household"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_importHouseholdDisaggregation"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "summary": "Invalid parameters"
          },
          "default": {
            "summary": "Error",
            "description": "Any other error, as RFC 9457 problem details",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/demo/household": {
      "get": {
        "tags": [
          "demo-household"
        ],
        "summary": "Household and model",
        "description": "Profile of the demo household (dataset, period, appliances, suggested window) and of the models that produced its results. Works without the database.",
        "operationId": "getDemoHousehold",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdOut"
                },
                "examples": {
                  "plegma101": {
                    "summary": "Plegma House 101",
                    "dataValue": {
                      "household": {
                        "name": "Plegma House 101",
                        "dataset": "Plegma",
                        "country": "GR",
                        "dataset_license": "CC BY 4.0",
                        "period": {
                          "start": "2022-07-15T00:00:00Z",
                          "end": "2023-09-30T22:23:00Z"
                        },
                        "timezone_note": "timestamps as recorded by the dataset, served as UTC",
                        "sampling_interval_seconds": 60,
                        "valid_days": 426.9,
                        "aggregate_energy_kwh": 3135.2,
                        "held_out": true,
                        "appliances": [
                          {
                            "key": "AC",
                            "name": "Điều hoà",
                            "name_en": "Air conditioner",
                            "kind": "appliance"
                          },
                          {
                            "key": "WaterHeater",
                            "name": "Bình nóng lạnh",
                            "name_en": "Water heater",
                            "kind": "appliance"
                          },
                          {
                            "key": "Fridge",
                            "name": "Tủ lạnh",
                            "name_en": "Fridge",
                            "kind": "appliance"
                          },
                          {
                            "key": "WashingMachine",
                            "name": "Máy giặt",
                            "name_en": "Washing machine",
                            "kind": "appliance"
                          },
                          {
                            "key": "Other",
                            "name": "Khác",
                            "name_en": "Other",
                            "kind": "residual"
                          }
                        ],
                        "sample_window": {
                          "start": "2023-07-10T00:00:00Z",
                          "end": "2023-07-17T00:00:00Z",
                          "bucket": "1 hour"
                        },
                        "missing_pct": 3.61,
                        "metered_pct": 81.01
                      },
                      "model": {
                        "name": "NILMFormer",
                        "parameters": 383283,
                        "window_minutes": 128,
                        "sampling": "1min",
                        "paper": "arXiv:2506.05880",
                        "git_commit": "a5936f9",
                        "per_appliance": [
                          {
                            "appliance": "AC",
                            "training_datasets": [
                              "Plegma",
                              "PRECON"
                            ],
                            "training_houses": 15,
                            "validation_house": 103,
                            "epochs": 15,
                            "best_epoch": 5
                          },
                          {
                            "appliance": "WaterHeater",
                            "training_datasets": [
                              "Plegma"
                            ],
                            "training_houses": 10,
                            "validation_house": 103,
                            "epochs": 22,
                            "best_epoch": 12
                          },
                          {
                            "appliance": "Fridge",
                            "training_datasets": [
                              "REFIT",
                              "Plegma"
                            ],
                            "training_houses": 18,
                            "validation_house": 103,
                            "epochs": 29,
                            "best_epoch": 19
                          },
                          {
                            "appliance": "WashingMachine",
                            "training_datasets": [
                              "REFIT",
                              "Plegma"
                            ],
                            "training_houses": 26,
                            "validation_house": 103,
                            "epochs": 20,
                            "best_epoch": 10
                          }
                        ]
                      }
                    }
                  }
                }
              }
            },
            "summary": "Household profile"
          },
          "404": {
            "description": "The household series is not imported, or the snapshot is missing.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "summary": "Demo data not available"
          },
          "default": {
            "summary": "Error",
            "description": "Any other error, as RFC 9457 problem details",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/demo/consumption": {
      "get": {
        "tags": [
          "demo-consumption"
        ],
        "summary": "Power per appliance over a window",
        "description": "Whole-house power and the **predicted** power of every appliance, bucketed, with the energy and share of each appliance. Without parameters: one summer week by hour. At most 2000 buckets per series.",
        "operationId": "getDemoConsumption",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Window start, ISO 8601 (UTC if no offset). Default: the sample window",
              "title": "Start"
            },
            "description": "Window start, ISO 8601 (UTC if no offset). Default: the sample window"
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Window end, exclusive. Default: the sample window",
              "title": "End"
            },
            "description": "Window end, exclusive. Default: the sample window"
          },
          {
            "name": "bucket",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "1 minute",
                "10 minutes",
                "1 hour",
                "1 day"
              ],
              "type": "string",
              "description": "Bucket width",
              "default": "1 hour",
              "title": "Bucket"
            },
            "description": "Bucket width"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsumptionOut"
                }
              }
            },
            "summary": "Bucketed series and totals"
          },
          "404": {
            "summary": "Demo data not available",
            "description": "The household series is not imported, or the snapshot is missing.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "summary": "Invalid parameters",
            "description": "Malformed value, `start` after `end`, or too many points.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          },
          "default": {
            "summary": "Error",
            "description": "Any other error, as RFC 9457 problem details",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/demo/breakdown": {
      "get": {
        "tags": [
          "demo-consumption"
        ],
        "summary": "Energy share per appliance",
        "description": "Predicted energy and share of each appliance over a window (for a pie or bar chart). Without parameters: the whole recorded period.",
        "operationId": "getDemoBreakdown",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Window start, ISO 8601 (UTC if no offset). Default: the sample window",
              "title": "Start"
            },
            "description": "Window start, ISO 8601 (UTC if no offset). Default: the sample window"
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Window end, exclusive. Default: the sample window",
              "title": "End"
            },
            "description": "Window end, exclusive. Default: the sample window"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BreakdownOut"
                }
              }
            },
            "summary": "Energy per appliance"
          },
          "404": {
            "summary": "Demo data not available",
            "description": "The household series is not imported, or the snapshot is missing.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "summary": "Invalid parameters",
            "description": "Malformed value, `start` after `end`, or too many points.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          },
          "default": {
            "summary": "Error",
            "description": "Any other error, as RFC 9457 problem details",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/demo/evaluation/monthly": {
      "get": {
        "tags": [
          "demo-evaluation"
        ],
        "summary": "Monthly energy, measured vs predicted",
        "operationId": "listDemoMonthlyEnergy",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonthlyOut"
                }
              }
            },
            "summary": "One item per month"
          },
          "404": {
            "description": "The household series is not imported, or the snapshot is missing.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "summary": "Demo data not available"
          },
          "default": {
            "summary": "Error",
            "description": "Any other error, as RFC 9457 problem details",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/demo/evaluation/daily": {
      "get": {
        "tags": [
          "demo-evaluation"
        ],
        "summary": "Daily energy, measured vs predicted",
        "description": "Days with at least 90 % of the minutes, optionally filtered by date.",
        "operationId": "listDemoDailyEnergy",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "First day (inclusive)",
              "title": "Start"
            },
            "description": "First day (inclusive)"
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Last day (inclusive)",
              "title": "End"
            },
            "description": "Last day (inclusive)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyOut"
                }
              }
            },
            "summary": "One item per day"
          },
          "404": {
            "summary": "Demo data not available",
            "description": "The household series is not imported, or the snapshot is missing.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "summary": "Invalid parameters",
            "description": "Malformed value, `start` after `end`, or too many points.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          },
          "default": {
            "summary": "Error",
            "description": "Any other error, as RFC 9457 problem details",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/demo/evaluation/sample-day": {
      "get": {
        "tags": [
          "demo-evaluation"
        ],
        "summary": "One day, measured vs predicted power",
        "description": "10-minute means of one summer day for every appliance. `null` marks an interval without data (the household has a 2-hour gap that day).",
        "operationId": "getDemoSampleDay",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SampleDayOut"
                }
              }
            },
            "summary": "144 ten-minute points"
          },
          "404": {
            "description": "The household series is not imported, or the snapshot is missing.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "summary": "Demo data not available"
          },
          "default": {
            "summary": "Error",
            "description": "Any other error, as RFC 9457 problem details",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Body_importHouseholdDisaggregation": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "CSV from nilmformer-experiment outputs"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_importHouseholdDisaggregation"
      },
...
      "ImportResult": {
        "properties": {
          "household_id": {
            "type": "string",
            "format": "uuid",
            "title": "Household Id"
          },
          "rows_per_device": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Rows Per Device"
          }
        },
        "type": "object",
        "required": [
          "household_id",
          "rows_per_device"
        ],
        "title": "ImportResult"
      },
...
      "Problem": {
        "additionalProperties": true,
        "description": "RFC 9457 problem details. Extension members may be added.",
        "properties": {
          "type": {
            "default": "about:blank",
            "description": "URI reference identifying the problem type",
            "title": "Type",
            "type": "string"
          },
          "title": {
            "description": "Short, human-readable summary of the problem type",
            "examples": [
              "Not Found"
            ],
            "title": "Title",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code",
            "examples": [
              404
            ],
            "title": "Status",
            "type": "integer"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Explanation specific to this occurrence",
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "URI reference of this occurrence (the path)",
            "title": "Instance"
          }
        },
        "required": [
          "title",
          "status"
        ],
        "title": "Problem",
        "type": "object"
      },
...
    }
  },
...
}